refactored code to language utils
